Refactor code: Move getNumberWithCommas function to util.ts and update import path for QuotationListItem
