+++ src/app/main/pricing/online-quote/list.tsx
@@ -12,7 +12,7 @@ import {
 } from "@/app/util/md3";
 import { Download, SwapHorizOutlined } from "@mui/icons-material";
 import NAOutlinedListBox from "@/app/components/na-outline-listbox";
-import { QuotationListItem } from "./listitem";
+import { QuotationListItem } from "./components/listitem";
 import { faker } from "@faker-js/faker";
 import { DateTime } from "luxon";
 import { createDummyVesselInformation } from "../../schedule/util";
